+++ R/GetRestCall.R
@@ -41,5 +41,11 @@ GetRestCall <- function(url) {
 	  HandleHTTPError(status, response)
 	}
 
-	RJSONIO::fromJSON(response)
+	# Prevent error on blank response
+	if(response == ""){
+	  invisible("")
+	} else {
+	  RJSONIO::fromJSON(response)
+	}
+	
 }