+++ NEWS.md
@@ -5,6 +5,7 @@
 - `GetMetadata` provides better error message when API gives truncated data
 - `GetHistory` now works where there is no Value
 - Fixed bug in `GetData` where it would fail if there was pivoting
+- No REST calls fail on blank now
 
 
 # CHANGES IN faosws VERSION 0.8.1